fix ovewrite config

+++ PKGBUILD
@@ -1,7 +1,7 @@
 # Maintainer: Mohamed Elsayed <mohamed@devlix.org>
 pkgname=smbmnt
 pkgver=r8.42f5345
-pkgrel=4
+pkgrel=5
 pkgdesc="Mount samba shares on Linux easily."
 arch=("x86_64")
 url="https://github.com/Mohamed1242012/smbmnt"
@@ -25,7 +25,9 @@ package() {
 	cd "$pkgname"
 	mkdir -p ~/.config/smbmnt
 	install -Dm755 ./smbmnt "$pkgdir/usr/bin/smbmnt"
-	install -Dm755 ./config.yaml "$HOME/.config/smbmnt/config.yaml"
+	if [ ! -f "$HOME/.config/smbmnt/config.yaml" ]; then
+    install -Dm755 ./config.yaml "$HOME/.config/smbmnt/config.yaml"
+	fi
 	install -Dm644 ./README.md "pkgdir/usr/share/doc/smbmnt/README.md"
   install -Dm644 ./LICENSE "$pkgdir/usr/share/licenses/smbmnt/LICENSE"
 }
